backend/requirements.txt
appnope==0.1.4
asttokens==3.0.0
backcall==0.2.0
bcrypt==4.3.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
exceptiongroup==1.2.2
executing==2.2.0
Flask==3.1.0
fonttools==4.57.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.7.5
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.24.4
packaging==24.2
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.50
psutil==7.0.0
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.1.4
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyzmq==26.4.0
seaborn==0.13.2
six==1.16.0
SQLAlchemy==2.0.30
stack-data==0.6.3
tomli==2.2.1
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.1
tzdata==2025.2
wcwidth==0.2.13
Werkzeug==3.1.3
zipp==3.20.2
